+++ src/app/auth/forgot-password/forgot-password.component.spec.ts
@@ -1,19 +1,19 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { CounterDashboardComponent } from "./counter-dashboard.component";
+import { ForgotPasswordComponent } from "./forgot-password.component";
 
-describe("CounterDashboardComponent", () => {
-  let component: CounterDashboardComponent;
-  let fixture: ComponentFixture<CounterDashboardComponent>;
+describe("ForgotPasswordComponent", () => {
+  let component: ForgotPasswordComponent;
+  let fixture: ComponentFixture<ForgotPasswordComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [CounterDashboardComponent]
+      declarations: [ForgotPasswordComponent]
     }).compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CounterDashboardComponent);
+    fixture = TestBed.createComponent(ForgotPasswordComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });