+++ src/app/auth/sign-in/sign-in.component.spec.ts
@@ -1,19 +1,19 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { CounterDashboardComponent } from "./counter-dashboard.component";
+import { SignInComponent } from "./sign-in.component";
 
-describe("CounterDashboardComponent", () => {
-  let component: CounterDashboardComponent;
-  let fixture: ComponentFixture<CounterDashboardComponent>;
+describe("SignInComponent", () => {
+  let component: SignInComponent;
+  let fixture: ComponentFixture<SignInComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [CounterDashboardComponent]
+      declarations: [SignInComponent]
     }).compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CounterDashboardComponent);
+    fixture = TestBed.createComponent(SignInComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });