+++ src/app/counter-dashboard/counter-dashboard.module.ts
@@ -2,20 +2,34 @@ import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { CounterComponent } from "./counter/counter.component";
 import { CounterDashboardComponent } from "./counter-dashboard.component";
-import { CounterService } from "./counter-service/counter.service";
+import { CounterService } from "./../shared/services/counter-service/counter.service";
 import { FormsModule } from "@angular/forms";
 import { FlexLayoutModule } from "@angular/flex-layout";
 import { MaterialModule } from "../material.module";
 import { CreateCounterDialogComponent } from "./create-counter-dialog/create-counter-dialog.component";
+import { ToolbarComponent } from "./toolbar/toolbar.component";
+import { AppRoutingModule } from "../app-routing.module";
 
 @NgModule({
   declarations: [
     CounterComponent,
     CounterDashboardComponent,
-    CreateCounterDialogComponent
+    CreateCounterDialogComponent,
+    ToolbarComponent
   ],
-  imports: [CommonModule, FlexLayoutModule, FormsModule, MaterialModule],
-  exports: [CounterComponent, CounterDashboardComponent],
+  imports: [
+    AppRoutingModule,
+    CommonModule,
+    FlexLayoutModule,
+    FormsModule,
+    MaterialModule
+  ],
+  exports: [
+    CounterComponent,
+    CounterDashboardComponent,
+    CreateCounterDialogComponent,
+    ToolbarComponent
+  ],
   providers: [CounterService],
   entryComponents: [CreateCounterDialogComponent]
 })